Support creating SVG nodes.
SVG nodes must be constructed with virtual-dom/virtual-hyperscript/svg rather than virtual-dom/h due to differences in how properties are handled for SVG nodes (see https://github.com/Matt-Esch/virtual-dom/issues/55).

Rather than listing all tag names that should be constructed with the SVG constructor (which would be problematic for the 'a' tag name, which is both an HTML and a SVG tag name), V assumes HTML tags, and when an SVG tag is created, it constructs in with the SVG constructor and recursively maps its children to using the SVG constructor, stopping recursion at any foreignObject children.